Initial code
Made some initial changes to the Longbow core code to introduce the
necessary hooks into the submit script and submit methods to override
them.

Regex still needs adding to bring back the jobid and also support for
subqueueing could be added to chemshell so that longbow can pick up the
error message.